Update gitlint and unit tests for windows
* Update config and hooks
* Update lots of tests

All the changes were to do with forward vs backward slashes.
Use os.path.join everywhere it's needed and add some re.escape()
in to deal with the cases where a backslash formatted path gets
put into a regexp. An abspath or two also required to make it all
work on both platforms.